VMS-14903: Server crashes on stop if ONVIF input port monitoring was enabled
<==== Mercurial data ====>
Branch: vms_4.0_server

+++ vms/server/nx_vms_server/src/plugins/resource/onvif/onvif_resource.cpp
@@ -3801,6 +3801,9 @@ void QnPlOnvifResource::stopInputPortStatesMonitoring()
         }
     }
 
+    if (serverModule()->isStopping())
+        return;
+
     if (QnSoapServer::instance() && QnSoapServer::instance()->getService())
         QnSoapServer::instance()->getService()->removeResourceRegistration(toSharedPointer(this));
